+++ src/main/java/com/haru/api/global/interceptor/DocumentMemberAuthInterceptor.java
@@ -1,7 +1,7 @@
 package com.haru.api.global.interceptor;
 
 import com.haru.api.workspace.domain.enums.DocumentType;
-import com.haru.api.domain.meeting.repository.MeetingRepository;
+import com.haru.api.meeting.infrastructure.MeetingRepository;
 import com.haru.api.domain.moodTracker.repository.MoodTrackerRepository;
 import com.haru.api.domain.snsEvent.repository.SnsEventRepository;
 import com.haru.api.user.domain.User;
