+++ src/main/java/com/haru/api/meeting/application/port/in/MeetingCommandUseCase.java
@@ -1,13 +1,13 @@
-package com.haru.api.domain.meeting.service;
+package com.haru.api.meeting.application.port.in;
 
-import com.haru.api.domain.meeting.dto.MeetingRequestDTO;
-import com.haru.api.domain.meeting.dto.MeetingResponseDTO;
-import com.haru.api.domain.meeting.entity.Meeting;
+import com.haru.api.meeting.presentation.dto.MeetingRequestDTO;
+import com.haru.api.meeting.presentation.dto.MeetingResponseDTO;
+import com.haru.api.meeting.domain.Meeting;
 import com.haru.api.user.domain.User;
 import com.haru.api.infra.websocket.AudioSessionBuffer;
 import org.springframework.web.multipart.MultipartFile;
 
-public interface MeetingCommandService {
+public interface MeetingCommandUseCase {
 
     MeetingResponseDTO.createMeetingResponse createMeeting(User user, MultipartFile agendaFile, MeetingRequestDTO.createMeetingRequest request);
