+++ src/main/java/com/haru/api/meeting/domain/Meeting.java
@@ -1,4 +1,4 @@
-package com.haru.api.domain.meeting.entity;
+package com.haru.api.meeting.domain;
 
 import com.haru.api.workspace.domain.Documentable;
 import com.haru.api.workspace.domain.enums.DocumentType;
