+++ src/main/java/com/haru/api/meeting/infrastructure/MeetingKeywordRepository.java
@@ -1,6 +1,6 @@
-package com.haru.api.domain.meeting.repository;
+package com.haru.api.meeting.infrastructure;
 
-import com.haru.api.domain.meeting.entity.MeetingKeyword;
+import com.haru.api.meeting.domain.MeetingKeyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MeetingKeywordRepository extends JpaRepository<MeetingKeyword, Long> {
