+++ src/main/java/com/haru/api/meeting/presentation/dto/MeetingResponseDTO.java
@@ -1,4 +1,4 @@
-package com.haru.api.domain.meeting.dto;
+package com.haru.api.meeting.presentation.dto;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
