+++ src/main/java/com/haru/api/workspace/application/converter/WorkspaceConverter.java
@@ -2,7 +2,7 @@ package com.haru.api.workspace.application.converter;
 
 import com.haru.api.workspace.domain.UserDocumentLastOpened;
 import com.haru.api.workspace.domain.enums.DocumentType;
-import com.haru.api.domain.meeting.entity.Meeting;
+import com.haru.api.meeting.domain.Meeting;
 import com.haru.api.domain.moodTracker.entity.MoodTracker;
 import com.haru.api.domain.snsEvent.entity.SnsEvent;
 import com.haru.api.user.presentation.dto.UserResponseDTO;
