+++ src/main/java/com/haru/api/workspace/application/service/WorkspaceQueryUseCaseImpl.java
@@ -3,8 +3,8 @@ package com.haru.api.workspace.application.service;
 import com.haru.api.workspace.application.port.in.WorkspaceQueryUseCase;
 import com.haru.api.workspace.domain.UserDocumentLastOpened;
 import com.haru.api.workspace.infrastructure.UserDocumentLastOpenedRepository;
-import com.haru.api.domain.meeting.entity.Meeting;
-import com.haru.api.domain.meeting.repository.MeetingRepository;
+import com.haru.api.meeting.domain.Meeting;
+import com.haru.api.meeting.infrastructure.MeetingRepository;
 import com.haru.api.domain.moodTracker.entity.MoodTracker;
 import com.haru.api.domain.moodTracker.repository.MoodTrackerRepository;
 import com.haru.api.domain.snsEvent.entity.SnsEvent;
