+++ src/main/java/com/haru/api/workspace/domain/Workspace.java
@@ -1,6 +1,6 @@
 package com.haru.api.workspace.domain;
 
-import com.haru.api.domain.meeting.entity.Meeting;
+import com.haru.api.meeting.domain.Meeting;
 import com.haru.api.domain.moodTracker.entity.MoodTracker;
 import com.haru.api.domain.snsEvent.entity.SnsEvent;
 import com.haru.api.global.common.entity.BaseEntity;
